+++ LocalNuGetHelper/Private/ExtractCsprojMetaData.ps1
@@ -1,5 +1,5 @@
 function ExtractCsprojMetaData {
-    Write-Verbose "Getting csproj of ${PSScriptRoot} and extract package name and version metadata."
+    Write-Verbose "Getting csproj of $PackageProjectPath and extract package name and version metadata."
     $csprojPath = Join-Path $PackageProjectPath "*.csproj"
     $csprojXml = [xml](Get-Content $csprojPath -ErrorAction Stop)