+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 \-
 
+## [0.3.2] - 2021-08-09
+
+### Fixed
+
+-   Wrong csproj directory path in verbose message
+-   Exclude filter of watch mode if the current location and the package project path are the same
+
 ## [0.3.1] - 2021-08-09
 
 ### Fixed